requirements.txt
fastapi
uvicorn
pymongo
pfmongo

pydantic[email]
pydantic_settings

pudb
toml
platformdirs
loguru

PyGithub
GitPython

